Make sidebar header configurable
